fhir-server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
fhir.resources==7.0.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-dotenv==1.0.0
python-multipart==0.0.6
